\name{fast_lm-package}
\alias{fast_lm-package}
\alias{fast_lm}
\docType{package}
\title{
\packageTitle{fast_lm}
}
\description{
\packageDescription{fast_lm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fast_lm}
\packageIndices{fast_lm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fast_lm}

Maintainer: \packageMaintainer{fast_lm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
